requirements.txt
pillow
pytesseract
